+++ src/core/component/directives/attrs/test/unit/main.ts
@@ -6,28 +6,27 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import type { ElementHandle, Locator, Page } from 'playwright';
-
-import type { Watcher } from 'components/directives/on-resize';
-
 import test from 'tests/config/unit/test';
 
-import { Component } from 'tests/helpers';
+import {
+
+	renderDirective,
+	renderDummy,
+
+	resizeHandler,
+	clickHandler,
+
+	waitForWatcherCallsCount,
+	dummyDeleteHandler
+
+} from 'core/component/directives/attrs/test/helpers';
 
 test.describe('core/component/directives/attrs', () => {
 	test.beforeEach(async ({demoPage}) => {
 		await demoPage.goto();
 	});
 
-	test('the directive allows setting the value of another directive', async ({page}) => {
-		const component = await renderDirective(page, 'b-dummy', {
-			'v-show': false
-		});
-
-		await test.expect(component).toHaveCSS('display', 'none');
-	});
-
-	test('the directive allows setting regular props or attributes', async ({page}) => {
+	test('should allow setting regular props or attributes', async ({page}) => {
 		const component = await renderDirective(page, 'b-dummy', {
 			style: 'margin-top: 10px;',
 			class: 'croatoan'
@@ -37,7 +36,7 @@ test.describe('core/component/directives/attrs', () => {
 		await test.expect(component).toHaveCSS('margin-top', '10px');
 	});
 
-	test('the directive allows setting event listeners with support of Vue modifiers', async ({page}) => {
+	test('should allow setting event listeners with support of Vue modifiers', async ({page}) => {
 		const component = await renderDirective(page, 'b-dummy', {
 			'@click.once': clickHandler
 		});
@@ -47,10 +46,17 @@ test.describe('core/component/directives/attrs', () => {
 		await component.click();
 
 		await test.expect(component).toHaveAttribute('data-counter', '1');
-
 	});
 
-	test('the directive allows setting custom directives', async ({page}) => {
+	test('should allow setting Vue directives', async ({page}) => {
+		const component = await renderDirective(page, 'b-dummy', {
+			'v-show': false
+		});
+
+		await test.expect(component).toHaveCSS('display', 'none');
+	});
+
+	test('should allow setting custom directives', async ({page}) => {
 		const component = await renderDirective(page, 'b-dummy', {
 			'v-on-resize': resizeHandler
 		});
@@ -62,7 +68,7 @@ test.describe('core/component/directives/attrs', () => {
 		await test.expect(waitForWatcherCallsCount(page, component, 2)).toBeResolved();
 	});
 
-	test('the directive allows specifying directives, events, and attributes simultaneously', async ({page}) => {
+	test('should allow specifying directives, events, and attributes simultaneously', async ({page}) => {
 		const component = await renderDirective(page, 'b-dummy', {
 			style: 'margin-top: 10px;',
 			'@click.once': clickHandler,
@@ -82,7 +88,7 @@ test.describe('core/component/directives/attrs', () => {
 		await test.expect(waitForWatcherCallsCount(page, component, 3)).toBeResolved();
 	});
 
-	test('the directive works correctly when used on functional components', async ({page}) => {
+	test('should allow specifying directives, events, and attributes simultaneously for functional components', async ({page}) => {
 		const component = await renderDirective(page, 'b-dummy-functional', {
 			style: 'margin-top: 10px;',
 			'@click.once': clickHandler,
@@ -101,56 +107,38 @@ test.describe('core/component/directives/attrs', () => {
 		await test.expect(component).toHaveAttribute('data-counter', '3');
 	});
 
-	async function renderDirective(
-		page: Page,
-		componentName: string,
-		attrs: RenderComponentsVnodeParams['attrs']
-	): Promise<Locator> {
-		const componentTestId = 'target';
-		await Component.createComponent(page, componentName, {
-			'data-testid': componentTestId,
-			'data-counter': 0,
-			'v-attrs': {...attrs},
-			style: 'width: 100px; height: 100px'
+	test('should work correctly with an event named as a registered key modifier', async ({page}) => {
+		const target = await renderDummy(page, {
+			'@delete': dummyDeleteHandler
 		});
 
-		return page.getByTestId(componentTestId);
-	}
+		const button = page.getByTestId('deleteButton');
 
-	async function waitForWatcherCallsCount(page: Page, observedEl: Locator, expected: number): Promise<void> {
-		const handle = await observedEl.elementHandle();
+		await button.click();
 
-		await page
-			.waitForFunction(
-				([div, val]) => Boolean(
-					div.getAttribute('data-counter') === val.toString(10)
-				),
+		await button.click();
 
-				<[ElementHandle<HTMLElement>, number]>[handle, expected]
-			);
-	}
+		await test.expect(target.evaluate((el) => el.counter)).resolves.toBe(2);
+	});
 
-	function resizeHandler(newRect: DOMRect, oldRect: DOMRect, watcher: Watcher): void {
-		const {target} = watcher;
+	test(
+		[
+			'should correctly handle registered keys modifiers:',
+			'our handler should be replaced by the registered handler'
+		].join(' '),
 
-		const previousValue = parseInt(
-			target.getAttribute('data-counter') ?? '0',
-			10
-		);
+		async ({page}) => {
+			const target = await renderDummy(page, {
+				'@delete.delete': dummyDeleteHandler
+			});
 
-		const nextValue = previousValue + 1;
-		target.setAttribute('data-counter', nextValue.toString());
-	}
+			const button = page.getByTestId('deleteButton');
 
-	function clickHandler(event: MouseEvent): void {
-		const target = <Element>event.target;
+			await button.click();
 
-		const previousValue = parseInt(
-			target.getAttribute('data-counter') ?? '0',
-			10
-		);
+			await button.click();
 
-		const nextValue = previousValue + 1;
-		target.setAttribute('data-counter', nextValue.toString());
-	}
+			await test.expect(target.evaluate((el) => el.counter)).resolves.toBe(0);
+		}
+	);
 });