+++ src/core/page-meta-data/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.110 (2024-07-17)
+
+#### :bug: Bug Fix
+
+* The page description element is now expected to be a meta tag
+  with the attribute `name='description'` instead of the `description` tag
+
 ## v4.0.0-beta.108.a-new-hope (2024-07-15)
 
 #### :boom: Breaking Change